notion-automator/daily-task-rotator/pyproject.toml
[tool.poetry]
name = "daily-task-rotator"
version = "0.1.0"
description = ""
authors = ["Aveek Das <aveek.22@gmail.com>"]
readme = "README.md"
packages = [{include = "daily_task_rotator"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
urllib3 = "1.26.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
